+++ mazarin/mancini/clockface_digit.go
@@ -45,7 +45,7 @@ func (f *DigitFace) DrawFace(dc DrawContext, fc *FontConfig, pal Palette, cx, cy
 
 	// Rotated digit markers at each hour position.
 	// Each digit is rotated by its angular position: 12→0°, 1→30°, 2→60°, ...
-	fontSize := math.Max(6, radius*0.20)
+	fontSize := int64(math.Max(6, radius*0.20))
 	loadFont(fc, dc, true, fontSize)
 	dc.SetColor(col)
 	markerRad := radius * 0.80
