+++ mazarin/mancini/font_config.go
@@ -8,7 +8,7 @@ import (
 // independent of palette or rendering configuration.
 type FontConfig struct {
 	// LoadFace creates a font.Face at the given size.
-	LoadFace func(bold bool, size float64) font.Face
+	LoadFace func(bold bool, size int64) font.Face
 
 	// FontRegular is the path to the regular-weight font file.
 	// Used when LoadFace is nil.
@@ -20,13 +20,13 @@ type FontConfig struct {
 }
 
 // MeasureText returns the advance width of text at the given font size.
-func (fc *FontConfig) MeasureText(text string, bold bool, fontSize float64) float64 {
+func (fc *FontConfig) MeasureText(text string, bold bool, fontSize int64) float64 {
 	if fc == nil || fc.LoadFace == nil {
-		return float64(len(text)) * fontSize * 0.6 // rough estimate
+		return float64(len(text)) * float64(fontSize) * 0.6 // rough estimate
 	}
 	face := fc.LoadFace(bold, fontSize)
 	if face == nil {
-		return float64(len(text)) * fontSize * 0.6
+		return float64(len(text)) * float64(fontSize) * 0.6
 	}
 	advance := font.MeasureString(face, text)
 	return float64(advance) / 64.0
